+++ src/featured/notices/services/notices.service.ts
@@ -1,6 +1,5 @@
 import { serverApi } from '@/shared/lib/api'
-import { timeAgo } from '@/shared/lib/utils/timeAgo'
-import type { ApiNotice, Notice, NoticeListResponse, CreateNoticeRequest, CreateNoticeResponse, UpdateNoticeRequest, UpdateNoticeResponse } from '@/featured/notices/types'
+import type { NoticeListResponse, CreateNoticeRequest, CreateNoticeResponse, UpdateNoticeRequest, UpdateNoticeResponse } from '@/featured/notices/types'
 
 export async function getNotices() {
   return serverApi.get<NoticeListResponse>('/api/v1/notices')
@@ -12,15 +11,4 @@ export async function createNotice(body: CreateNoticeRequest) {
 
 export async function updateNotice(id: number, body: UpdateNoticeRequest) {
   return serverApi.patch<UpdateNoticeResponse>(`/api/v1/notices/${id}`, body)
-}
-
-export function mapApiNoticeToNotice(notice: ApiNotice): Notice {
-  return {
-    id: String(notice.id),
-    title: notice.title,
-    content: notice.content,
-    isActive: true,
-    createdAt: timeAgo(notice.createdAt),
-    updatedAt: timeAgo(notice.updatedAt),
-  }
 }
