+++ src/featured/questions/services/questions.service.ts
@@ -1,8 +1,5 @@
 import { serverApi } from '@/shared/lib/api'
-import { timeAgo } from '@/shared/lib/utils/timeAgo'
 import type {
-  ApiQuestion,
-  CommonQuestion,
   CreateQuestionRequest,
   CreateQuestionResponse,
   DeleteQuestionResponse,
@@ -25,16 +22,4 @@ export async function updateQuestion(id: string, body: UpdateQuestionRequest) {
 
 export async function deleteQuestion(id: string) {
   return serverApi.delete<DeleteQuestionResponse>(`/api/v1/questions/${id}`)
-}
-
-export function mapApiQuestionToCommon(question: ApiQuestion): CommonQuestion {
-  return {
-    id: question.id,
-    content: question.content,
-    category: '자기소개',
-    isActive: question.status === 'ACTIVE',
-    usageCount: 0,
-    createdAt: timeAgo(question.createdAt),
-    updatedAt: timeAgo(question.updatedAt),
-  }
 }
